Add and use gerrit username from config file
Using current process' username as gerrit username for connecting to
gerrit via SSH doesn't work as it usually is the apache2 user www-data
which doesn't have SSH public key for use with gerrit server.

Add a config line for username to fetch configuration file and use
username from config file when connecting to gerrit. Username in config
file has to be the same as username in SSH public key installed on
gerrit server.